requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0

# Optional dependencies for enhanced features
colorama>=0.4.6  # For colored terminal output on Windows
python-dotenv>=1.0.0  # For .env file support

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
